lib/api.ts 削除
