+++ src/lib/fileSystem.ts
@@ -1,5 +1,7 @@
-import { existsSync, readdirSync, statSync } from "fs";
 import path from "path";
+import fs from "fs";
+import { join } from "path";
+import matter from "gray-matter";
 
 // _notesフォルダの場所(constant.tsに置くとstorybookでエラー起きるのでここに配置する)
 export const NOTES_DIR = path.join(process.cwd(), "_notes");
@@ -8,7 +10,7 @@ export const NOTES_DIR = path.join(process.cwd(), "_notes");
 export function isDirectory(path: string) {
   // ディレクトリか判定
   try {
-    if (existsSync(path) && statSync(path).isDirectory()) {
+    if (fs.existsSync(path) && fs.statSync(path).isDirectory()) {
       return true;
     }
   } catch (error) {
@@ -16,7 +18,7 @@ export function isDirectory(path: string) {
   } finally {
     // .mdファイルであるか判定
     const mdPath = path + ".md";
-    if (existsSync(mdPath)) {
+    if (fs.existsSync(mdPath)) {
       return false;
     }
     // それでも存在しない → エラーが投げられる
@@ -35,7 +37,7 @@ export function getNoteUnderDirSlugs(
   includeMySelf?: boolean
 ) {
   // rootDirectory直下のファイル・ディレクトリ取得
-  const rootEnts = readdirSync(rootDirectory, { withFileTypes: true });
+  const rootEnts = fs.readdirSync(rootDirectory, { withFileTypes: true });
 
   const files: (FileSlugs | FileSlugs[])[] = [];
   for (const dirent of rootEnts) {
@@ -78,4 +80,58 @@ export function getNoteUnderDirSlugs(
     });
   }
   return files.flat();
+}
+
+const postsDirectory = join(process.cwd(), "_posts");
+
+// _postsフォルダ以下全ファイルのpath取得
+export function getPostSlugs() {
+  return fs.readdirSync(postsDirectory);
+}
+
+export function getPostBySlug(slug: string, fields: string[] = []) {
+  const realSlug = slug.replace(/\.md$/, "");
+  const fullPath = join(postsDirectory, `${realSlug}.md`);
+  const fileContents = fs.readFileSync(fullPath, "utf8");
+  const { data, content } = matter(fileContents);
+
+  type Items = {
+    [key: string]: string;
+  };
+
+  const items: Items = {};
+
+  // .mdでdraft: trueとなっているものは作成しない（非表示）
+  if (data["draft"] === true) {
+    return null;
+  }
+
+  // Ensure only the minimal needed data is exposed
+  fields.forEach((field) => {
+    if (field === "slug") {
+      items[field] = realSlug;
+    }
+    if (field === "content") {
+      items[field] = content;
+    }
+
+    if (typeof data[field] !== "undefined") {
+      items[field] = data[field];
+    }
+  });
+
+  return items;
+}
+
+// postsトップページに表示する記事の詳細情報を取得
+export function getAllPosts(fields: string[] = []) {
+  const slugs = getPostSlugs();
+  const posts = slugs
+    .map((slug) => getPostBySlug(slug, fields))
+    .filter((item) => {
+      return item;
+    })
+    // sort posts by date in descending order
+    .sort((post1, post2) => (post1.date > post2.date ? -1 : 1));
+  return posts;
 }
